Create import function and data
